Moves publishTweet to its own module
